refactor: modularize start_training.py into run_tracker.py and checkpoint_utils.py
