+++ start_training.py
@@ -37,196 +37,12 @@ from schedules import (
 )
 
 from metrics_logger import LossMonitor, MetricsLogger
+from run_tracker import RunTracker
+from checkpoint_utils import discover_latest_checkpoint_run, load_or_create_checkpoint
 import json
 import datetime
 import subprocess
 import sys
-
-class RunTracker:
-    def __init__(self, runs_root="runs"):
-        self.runs_root = runs_root
-        self.run_id = None
-        self.run_dir = None
-        self.start_time = None
-        self.session_index = None
-
-    @staticmethod
-    def get_git_metadata():
-        metadata = {
-            "commit": "unknown",
-            "branch": "unknown",
-            "dirty": False
-        }
-        try:
-            commit = subprocess.check_output(["git", "rev-parse", "HEAD"], stderr=subprocess.DEVNULL).decode().strip()
-            metadata["commit"] = commit
-        except Exception:
-            pass
-
-        try:
-            branch = subprocess.check_output(["git", "rev-parse", "--abbrev-ref", "HEAD"], stderr=subprocess.DEVNULL).decode().strip()
-            metadata["branch"] = branch
-        except Exception:
-            pass
-
-        try:
-            status = subprocess.check_output(["git", "status", "--porcelain"], stderr=subprocess.DEVNULL).decode().strip()
-            metadata["dirty"] = len(status) > 0
-        except Exception:
-            pass
-
-        return metadata
-
-    @staticmethod
-    def get_hyperparameters():
-        return {
-            "LATENT_DIM": LATENT_DIM,
-            "NUM_BLOCKS": NUM_BLOCKS,
-            "SHARED_SLOTS": SHARED_SLOTS,
-            "MAX_SEQ_LEN": MAX_SEQ_LEN,
-            "VOCAB_SIZE": VOCAB_SIZE,
-            "MAX_STEPS_LIMIT": MAX_STEPS_LIMIT,
-            "BATCH_SIZE": BATCH_SIZE,
-            "ACCUMULATION_STEPS": ACCUMULATION_STEPS,
-            "PAD_TOKEN_ID": PAD_TOKEN_ID,
-            "NUM_HEADS": NUM_HEADS,
-            "NUM_GROUPS": NUM_GROUPS,
-        }
-
-    def _check_compatibility(self, metadata_path):
-        if not os.path.exists(metadata_path):
-            return
-        
-        try:
-            with open(metadata_path, "r") as f:
-                old_meta = json.load(f)
-            
-            old_params = old_meta.get("parameters", {})
-            current_params = self.get_hyperparameters()
-            
-            critical_keys = [
-                "LATENT_DIM", "NUM_BLOCKS", "SHARED_SLOTS", "MAX_SEQ_LEN", 
-                "VOCAB_SIZE", "NUM_HEADS", "NUM_GROUPS"
-            ]
-            mismatches = [
-                f"  - {k}: run used {old_params[k]}, current code uses {current_params[k]}"
-                for k in critical_keys
-                if k in old_params and old_params[k] != current_params[k]
-            ]
-            
-            if mismatches:
-                print("\n" + "🛑"*20)
-                print("🛑 ERROR: Parameter Mismatch Detected! Cannot resume this training run:")
-                print("\n".join(mismatches))
-                print("\n💡 Options:")
-                print("  1. Revert your code parameters back to match the run's parameters.")
-                print("  2. Start a brand new training run with: python start_training.py --new-run")
-                print("  3. Point to a different checkpoint folder with: python start_training.py --checkpoint-path <path>")
-                print("🛑"*20 + "\n")
-                sys.exit(1)
-        except SystemExit:
-            sys.exit(1)
-        except Exception:
-            pass
-
-    def start_session(self, run_id=None):
-        os.makedirs(self.runs_root, exist_ok=True)
-        self.start_time = time.time()
-        start_timestamp = datetime.datetime.now().astimezone().isoformat()
-
-        if run_id is None:
-            # Generate a new unique run ID
-            timestamp_str = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
-            self.run_id = f"run_{timestamp_str}"
-            self.run_dir = os.path.join(self.runs_root, self.run_id)
-            os.makedirs(self.run_dir, exist_ok=True)
-
-            git_meta = self.get_git_metadata()
-            params = self.get_hyperparameters()
-
-            metadata = {
-                "run_id": self.run_id,
-                "git_commit": git_meta["commit"],
-                "git_branch": git_meta["branch"],
-                "git_dirty": git_meta["dirty"],
-                "parameters": params,
-                "sections": [
-                    {
-                        "start_time": start_timestamp,
-                        "end_time": None,
-                        "duration_seconds": None
-                    }
-                ]
-            }
-            self.session_index = 0
-            self.save_metadata(metadata)
-            print(f"📁 Created new training run folder: {self.run_dir}")
-        else:
-            # Resume existing run
-            self.run_id = run_id
-            self.run_dir = os.path.join(self.runs_root, self.run_id)
-            os.makedirs(self.run_dir, exist_ok=True)
-
-            metadata_path = os.path.join(self.run_dir, "run_metadata.json")
-            self._check_compatibility(metadata_path)
-
-            if os.path.exists(metadata_path):
-                try:
-                    with open(metadata_path, "r") as f:
-                        metadata = json.load(f)
-                except Exception:
-                    metadata = None
-            else:
-                metadata = None
-
-            if metadata is None:
-                git_meta = self.get_git_metadata()
-                params = self.get_hyperparameters()
-                metadata = {
-                    "run_id": self.run_id,
-                    "git_commit": git_meta["commit"],
-                    "git_branch": git_meta["branch"],
-                    "git_dirty": git_meta["dirty"],
-                    "parameters": params,
-                    "sections": []
-                }
-
-            metadata["sections"].append({
-                "start_time": start_timestamp,
-                "end_time": None,
-                "duration_seconds": None
-            })
-            self.session_index = len(metadata["sections"]) - 1
-            self.save_metadata(metadata)
-            print(f"🔄 Resumed training run folder: {self.run_dir}")
-
-        return self.run_id
-
-    def update_session_duration(self):
-        if self.run_dir is None or self.session_index is None:
-            return
-        metadata_path = os.path.join(self.run_dir, "run_metadata.json")
-        if not os.path.exists(metadata_path):
-            return
-        
-        try:
-            with open(metadata_path, "r") as f:
-                metadata = json.load(f)
-            
-            end_timestamp = datetime.datetime.now().astimezone().isoformat()
-            duration = time.time() - self.start_time
-            
-            metadata["sections"][self.session_index]["end_time"] = end_timestamp
-            metadata["sections"][self.session_index]["duration_seconds"] = round(duration, 2)
-            
-            self.save_metadata(metadata)
-        except Exception:
-            pass
-
-    def save_metadata(self, metadata):
-        metadata_path = os.path.join(self.run_dir, "run_metadata.json")
-        with open(metadata_path, "w") as f:
-            json.dump(metadata, f, indent=2)
 
 load_dotenv()
 
@@ -289,76 +105,6 @@ def create_sft_optimizer(model, old_state=None):
         
     gc.collect()
     return new_opt
-
-def discover_latest_checkpoint_run(runs_root="runs"):
-    if not os.path.exists(runs_root):
-        return None, None
-    
-    import glob
-    run_dirs = sorted(glob.glob(os.path.join(runs_root, "run_*")))
-    
-    for r_dir in reversed(run_dirs):
-        chk_dir = os.path.join(r_dir, "checkpoints")
-        if os.path.exists(chk_dir):
-            try:
-                mngr = ocp.CheckpointManager(
-                    chk_dir,
-                    item_names=("model", "optimizer", "monitor_state", "step"),
-                )
-                if mngr.latest_step() is not None:
-                    run_id = os.path.basename(r_dir)
-                    return chk_dir, run_id
-            except Exception:
-                pass
-    return None, None
-
-def load_or_create_checkpoint(model, optimizer, checkpoint_path, force_new_run=False):
-    monitor = LossMonitor()
-    mngr = ocp.CheckpointManager(
-        checkpoint_path,
-        item_names=("model", "optimizer", "monitor_state", "step"),
-        options=ocp.CheckpointManagerOptions(max_to_keep=3, create=True),
-    )
-
-    run_id = None
-    if not force_new_run and mngr.latest_step() is not None:
-        latest_step = mngr.latest_step()
-        print(f"📖 Loading Orbax checkpoint from step {latest_step}...")
-        restored = mngr.restore(
-            latest_step,
-            args=ocp.args.Composite(
-                model=ocp.args.StandardRestore(nnx.state(model)),
-                optimizer=ocp.args.StandardRestore(nnx.state(optimizer)),
-                monitor_state=ocp.args.JsonRestore(),
-                step=ocp.args.JsonRestore(),
-            ),
-        )
-
-        nnx.update(model, restored["model"])
-        nnx.update(optimizer, restored["optimizer"])
-        
-        start_step = restored["step"] + 1
-        m_state = restored["monitor_state"]
-        monitor.ce_history = m_state.get("ce_history", [])
-        monitor.best_ce = m_state.get("best_ce", float("inf"))
-        monitor.best_loss = m_state.get("best_loss", float("inf"))
-        monitor.best_avg_ce = m_state.get("best_avg_ce", monitor.best_ce)
-        monitor.last_improvement_step = m_state.get("last_improvement_step", 0)
-        monitor.sft_start_step = m_state.get("sft_start_step", None)
-        run_id = m_state.get("run_id", None)
-        
-        print(f"✅ Resuming from step {start_step}")
-        del restored 
-        import gc; gc.collect()
-    else:
-        if force_new_run:
-            print("🆕 Force New Run specified, starting from scratch...")
-        else:
-            print("🆕 No checkpoint found, starting from scratch...")
-        start_step = 1
-        monitor.sft_start_step = None
-
-    return mngr, monitor, start_step, optimizer, run_id
 
 def setup_data_pipeline(start_step, sft_phase_event, sft_start_step=None):
     print("🚀 Initializing Dynamic Data Phases...")
